==> preprocessing/3_transformation.ipynb <==
### 전처리 노트북 3/5: 변수 정규화
사용법: 
- 전처리 노트북 2를 실행 후 01_electric_train_features.csv, 01_electric_test_features.csv 를 asset_path에 저장 
- 만약 data_path가 없는 경로라면 생성해주세요
- 각 셀을 순서대로 실행

결과물:
- 02_train_input_transformed.csv
- 02_train_elec_transformed.csv
- 02_train_meta_transformed.csv
- 02_test_input_transformed.csv
- 02_test_meta_transformed.csv

# Standard library imports
from pathlib import Path
from datetime import datetime
import json
import pickle

from concurrent.futures import ProcessPoolExecutor, as_completed
from tqdm import tqdm

# Data handling and preprocessing
import pandas as pd
import numpy as np
from sklearn.preprocessing import QuantileTransformer, MinMaxScaler, StandardScaler, OneHotEncoder


# Parallel processing
from multiprocessing import Pool

# YAML file handling
import yaml

from preprocessing_utils import Backup, get_data_paths
asset_path, data_path = get_data_paths()

SAVE = True

train_data = pd.read_csv((Path(asset_path, '01_electric_train_features.csv')))
test_data  = pd.read_csv((Path(asset_path, '01_electric_test_features.csv')))

train_data['datetime'] = pd.to_datetime(train_data['datetime'])
test_data['datetime'] = pd.to_datetime(test_data['datetime'])

pre_num_train      = train_data['num'].values
pre_datetime_train = train_data['datetime'].values

pre_num_test      = test_data['num'].values
pre_datetime_test = test_data['datetime'].values

print(train_data.columns)
train_data.head()

print(test_data.columns)
test_data.head()

____

elec_variables = ['sum_qctr', 'sum_load', 'avg_load','n_mean_load', 'elec']

standard_transformed = ['temp', 'humid', 'tchi', 'dci', 'hi', 'wchi', 'atemp']
to_quantile_transformed = ['rain', 'units', 'sum_qctr', 'sum_load', 'avg_load','n_mean_load', 'elec']
min_max_transformed = ['lon', 'lat', 'wind', 'altitude']
categorical_to_onehot = [('district',[0,1,2,3])]
no_transform = ['is_dummy', 'num', 'year','datetime', 'holiday'] + elec_variables

input_cols = ['district_c0', 'district_c1', 'district_c2', 'district_c3', 'units_qt', 'lon_mm', 'lat_mm', 'holiday', 'altitude_mm', 'temp_st', 'humid_st', 'tchi_st', 'dci_st', 'hi_st', 'wchi_st', 'atemp_st', 'rain_qt', 'wind_mm']
elec_cols  = ['sum_qctr_qt', 'sum_load_qt', 'avg_load_qt', 'n_mean_load_qt', 'elec_qt', 'sum_qctr', 'sum_load', 'avg_load', 'n_mean_load', 'elec']
meta_cols  = ['is_dummy', 'num', 'year', 'datetime']

### Fitting Transforms on Train Data (Calculate Statistics etc..)

transforms = {}

train_value_dict = {}

for tag, cols in zip(['_st', '_qt', '_mm'], [standard_transformed, to_quantile_transformed, min_max_transformed]):
    for col in cols:

        col_values = train_data[[col]].values

        if col in standard_transformed:
            qt = StandardScaler()
        elif col in to_quantile_transformed:
            qt = QuantileTransformer(output_distribution='uniform')
        elif col in min_max_transformed:
            qt = MinMaxScaler()
        else:
            print('error')
            break
            
        col_transformed = qt.fit_transform(col_values)
        transforms[col] = qt

        train_value_dict[col + tag] = col_transformed.ravel()

        if col in elec_variables:
            with open(Path(data_path,f'{col}_transform.pkl'), 'wb') as f:
                pickle.dump(transforms[col], f)

for col, cat in categorical_to_onehot:
    col_values = train_data[col].values
    qt = OneHotEncoder(categories=[cat])
    col_transformed = qt.fit_transform(col_values.reshape(-1,1)).toarray()
    transforms[col] = qt
    for i in range(col_transformed.shape[1]):
        train_value_dict[col + f'_c{i}'] = col_transformed[:,i].ravel()

for col in no_transform:
    train_value_dict[col] = train_data[col].values.ravel()

train_value_dict.keys()

assert set(input_cols + elec_cols + meta_cols) == set(train_value_dict.keys())

### Apply Transforms on Test Data

test_value_dict = {}

for tag, cols in zip(['_st', '_qt', '_mm'], [standard_transformed, to_quantile_transformed, min_max_transformed]):
    for col in cols:
        if col in elec_variables: continue
        
        col_values = test_data[[col]].values

        qt = transforms[col]
        col_transformed = qt.transform(col_values)

        test_value_dict[col + tag] = col_transformed.ravel()

        if col in min_max_transformed:
            # trim the values to 0~1
            test_value_dict[col + tag] = np.clip(test_value_dict[col + tag], 0, 1).ravel()

for col, cat in categorical_to_onehot:
    
    qt = transforms[col]
    col_values = test_data[col].values
    col_transformed = qt.transform(col_values.reshape(-1,1)).toarray()
    for i in range(col_transformed.shape[1]):
        test_value_dict[col + f'_c{i}'] = col_transformed[:,i].ravel()

for col in no_transform:
    if col in elec_variables: continue
    test_value_dict[col] = test_data[col].values.ravel()


assert set(input_cols + meta_cols) == set(test_value_dict.keys()), print(set(input_cols + meta_cols) - set(test_value_dict.keys()))

train_whole_df = pd.DataFrame(train_value_dict)
test_whole_df = pd.DataFrame(test_value_dict)

train_input_df = train_whole_df[input_cols]
train_elec_df = train_whole_df[elec_cols]
train_meta_df = train_whole_df[meta_cols]

test_input_df = test_whole_df[input_cols]
test_meta_df = test_whole_df[meta_cols]

train_input_df.to_csv(Path(asset_path, '02_train_input_transformed.csv'), index=False)
train_elec_df.to_csv(Path(asset_path, '02_train_elec_transformed.csv'), index=False)
train_meta_df.to_csv(Path(asset_path, '02_train_meta_transformed.csv'), index=False)

test_input_df.to_csv(Path(asset_path, '02_test_input_transformed.csv'), index=False)
test_meta_df.to_csv(Path(asset_path, '02_test_meta_transformed.csv'), index=False)

___
